move channel search into ChannelContent component
