+++ ui/component/channelContent/view.jsx
@@ -1,11 +1,17 @@
 // @flow
+import * as CS from 'constants/claim_search';
+import * as ICONS from 'constants/icons';
 import React, { Fragment } from 'react';
 import HiddenNsfwClaims from 'component/hiddenNsfwClaims';
-import { withRouter } from 'react-router-dom';
+import { useHistory } from 'react-router-dom';
 import Button from 'component/button';
 import ClaimListDiscover from 'component/claimListDiscover';
-import * as CS from 'constants/claim_search';
 import Ads from 'web/component/ads';
+import Icon from 'component/common/icon';
+import { Form, FormField } from 'component/common/form';
+import { DEBOUNCE_WAIT_DURATION_MS } from 'constants/search';
+
+const LIGHTHOUSE_URL = 'https://lighthouse.lbry.com/search';
 
 type Props = {
   uri: string,
@@ -21,6 +27,7 @@ type Props = {
   defaultInfiniteScroll?: Boolean,
   claim: ?Claim,
   isAuthenticated: boolean,
+  showMature: boolean,
 };
 
 function ChannelContent(props: Props) {
@@ -34,8 +41,56 @@ function ChannelContent(props: Props) {
     isAuthenticated,
     defaultPageSize = CS.PAGE_SIZE,
     defaultInfiniteScroll = true,
+    showMature,
   } = props;
   const claimsInChannel = (claim && claim.meta.claims_in_channel) || 0;
+  const [searchQuery, setSearchQuery] = React.useState('');
+  const [searchResults, setSearchResults] = React.useState(undefined);
+  const {
+    location: { pathname, search },
+  } = useHistory();
+  const url = `${pathname}${search}`;
+  const claimId = claim && claim.claim_id;
+
+  function handleInputChange(e) {
+    const { value } = e.target;
+    setSearchQuery(value);
+  }
+
+  function getResults(fetchUrl) {
+    fetch(fetchUrl)
+      .then(res => res.json())
+      .then(results => {
+        const urls = results.map(({ name, claimId }) => {
+          return `lbry://${name}#${claimId}`;
+        });
+        setSearchResults(urls);
+      })
+      .catch(() => {
+        setSearchResults(null);
+      });
+  }
+
+  React.useEffect(() => {
+    const timer = setTimeout(() => {
+      if (searchQuery === '' || !claimId) {
+        // In order to display original search results, search results must be set to null. A query of '' should display original results.
+        return setSearchResults(null);
+      } else {
+        getResults(
+          `${LIGHTHOUSE_URL}?s=${encodeURIComponent(searchQuery)}&channel_id=${encodeURIComponent(claimId)}${
+            !showMature ? '&nsfw=false' : ''
+          }`
+        );
+      }
+    }, DEBOUNCE_WAIT_DURATION_MS);
+    return () => clearTimeout(timer);
+  }, [claimId, searchQuery, showMature]);
+
+  React.useEffect(() => {
+    setSearchQuery('');
+    setSearchResults(null);
+  }, [url]);
 
   return (
     <Fragment>
@@ -66,11 +121,24 @@ function ChannelContent(props: Props) {
 
       {claim && claimsInChannel > 0 ? (
         <ClaimListDiscover
+          uris={searchResults}
           channelIds={[claim.claim_id]}
           defaultOrderBy={CS.ORDER_BY_NEW}
           pageSize={defaultPageSize}
           infiniteScroll={defaultInfiniteScroll}
           injectedItem={!isAuthenticated && IS_WEB && <Ads type="video" />}
+          meta={
+            <Form onSubmit={() => {}} className="wunderbar--inline">
+              <Icon icon={ICONS.SEARCH} />
+              <FormField
+                className="wunderbar__input"
+                value={searchQuery}
+                onChange={handleInputChange}
+                type="text"
+                placeholder={__('Search')}
+              />
+            </Form>
+          }
         />
       ) : (
         <section className="main--empty">{__("This channel hasn't published anything yet")}</section>
@@ -79,4 +147,4 @@ function ChannelContent(props: Props) {
   );
 }
 
-export default withRouter(ChannelContent);
+export default ChannelContent;
