+++ ui/page/channel/index.js
@@ -1,5 +1,4 @@
 import { connect } from 'react-redux';
-import * as settings from 'constants/settings';
 import {
   makeSelectClaimIsMine,
   makeSelectTitleForUri,
@@ -11,7 +10,6 @@ import {
 } from 'lbry-redux';
 import { selectBlackListedOutpoints, doFetchSubCount, makeSelectSubCountForUri } from 'lbryinc';
 import { makeSelectIsSubscribed } from 'redux/selectors/subscriptions';
-import { makeSelectClientSetting } from 'redux/selectors/settings';
 import ChannelPage from './view';
 
 const select = (state, props) => ({
@@ -24,7 +22,6 @@ const select = (state, props) => ({
   isSubscribed: makeSelectIsSubscribed(props.uri, true)(state),
   channelIsBlocked: selectChannelIsBlocked(props.uri)(state),
   blackListedOutpoints: selectBlackListedOutpoints(state),
-  showMature: makeSelectClientSetting(settings.SHOW_MATURE)(state),
   subCount: makeSelectSubCountForUri(props.uri)(state),
 });
